requirements.txt
pytorch~=1.10.2
opencv-python~=4.3.0.38
numpy~=1.19.2
torchvision~=0.11.3
pillow~=8.4.0
scikit-image~=0.17.2